+++ src/app/projects/states/dashboard/directives/progress-dashboard/task-planner-card/task-planner-card.component.ts
@@ -1,0 +1,14 @@
+import {Component, Input} from '@angular/core';
+import {Project} from 'src/app/api/models/project';
+
+@Component({
+  selector: 'f-task-planner-card',
+  templateUrl: './task-planner-card.component.html',
+  styleUrl: './task-planner-card.component.scss',
+})
+export class TaskPlannerCardComponent {
+  @Input() project: Project;
+  public get unit() {
+    return this.project?.unit;
+  }
+}