+++ src/app/projects/states/plan/project-plan.component.ts
@@ -1,16 +1,44 @@
-import {Component} from '@angular/core';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {MatSelectChange} from '@angular/material/select';
+import {Project, ProjectService} from 'src/app/api/models/doubtfire-model';
+import {ConfirmationModalService} from 'src/app/common/modals/confirmation-modal/confirmation-modal.service';
+import {AlertService} from 'src/app/common/services/alert.service';
+import {GradeService} from 'src/app/common/services/grade.service';
 import {GlobalStateService} from '../index/global-state.service';
-import { Project, TaskDefinition } from 'src/app/api/models/doubtfire-model';
+import {TaskPlannerComponent} from './task-planner/task-planner.component';
 
 @Component({
   selector: 'f-project-plan',
   templateUrl: 'project-plan.component.html',
-  // styleUrls: ['project-plan.component.scss']
+  styleUrls: ['project-plan.component.scss'],
 })
-export class ProjectPlanComponent {
+export class ProjectPlanComponent implements OnInit {
   public project: Project;
 
-  constructor(private globalStateService: GlobalStateService) {
+  @ViewChild(TaskPlannerComponent) planner!: TaskPlannerComponent;
+
+  public get unit() {
+    return this.project?.unit;
+  }
+
+  public get gradeValues() {
+    return this.gradeService.gradeValues;
+  }
+
+  public get gradeAcronyms() {
+    return this.gradeService.gradeAcronyms;
+  }
+
+  public gradeString(grade: number) {
+    return this.gradeService.grades[grade];
+  }
+
+  constructor(
+    private globalStateService: GlobalStateService,
+    private gradeService: GradeService,
+    private projectService: ProjectService,
+    private alertService: AlertService,
+  ) {
     this.globalStateService.currentViewAndEntitySubject$.subscribe((viewAndEntity) => {
       if (viewAndEntity.viewType === 'PROJECT' && viewAndEntity.entity) {
         this.project = viewAndEntity.entity as Project;
@@ -18,13 +46,26 @@ export class ProjectPlanComponent {
     });
   }
 
-  public taskDefs(): TaskDefinition[] {
-    if (!this.project || !this.project.unit.taskDefinitions) {
-      return [];
-    }
+  public selectedTargetGrade: number;
 
-    return this.project.unit.taskDefinitions.filter((taskDef) => {
-      return taskDef.targetGrade <= this.project.targetGrade;
+  ngOnInit(): void {
+    this.selectedTargetGrade = this.project.targetGrade;
+  }
+
+  onTargetGradeChange(event: MatSelectChange) {
+    const previousTargetGrade = this.project.targetGrade;
+    this.project.targetGrade = event.value;
+
+    this.projectService.update(this.project).subscribe({
+      next: () => {
+        this.alertService.success(`Succesfully updated target grade`, 2000);
+        this.planner.refreshItems();
+      },
+      error: (error) => {
+        this.project.targetGrade = previousTargetGrade;
+        this.selectedTargetGrade = previousTargetGrade;
+        this.alertService.error(`Failed to update target grade: ${error}`, 6000);
+      },
     });
   }
 }